local-up-cluster: Allow to enable --allocate-node-cidrs option directly
This change introduces a new environment variable ALLOCATE_NODE_CIDRS
which allows to enable the --allocate-node-cidrs option in the
kube-controller-manager.

Previously, that option was enabled only when NET_PLUGIN="kubenet" was
set. To preserve that behavior, NET_PLUGIN="kubenet" triggers setting
ALLOCATE_NODE_CIDRS="true" even if that variable was not set.

The motivation behind allowing to set ALLOCATE_NODE_CIDRS directly is
ability to use local-up-cluster with network providers which rely on
PodCIDR allocation (i.e. Cilium).

Signed-off-by: Michal Rostecki <mrostecki@opensuse.org>